policy added

+++ app/Http/Controllers/ChirpController.php
@@ -67,6 +67,8 @@ class ChirpController extends Controller
      */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         // Validate
         $validated = $request->validate([
             'message' => 'required|string|max:255',
@@ -83,6 +85,8 @@ class ChirpController extends Controller
      */
     public function destroy(Chirp $chirp)
     {
+        $this->authorize('delete', $chirp);
+
         $chirp->delete();
     
         return redirect('/')->with('success', 'Chirp deleted!');
